requirements.txt
backports.ssl-match-hostname==3.4.0.2
certifi==2016.2.28
Django==1.8.12
tornado==4.3
redis==2.10.5
tornado-redis==2.4.18
django-registration==2.0.4
python-social-auth==0.2.18
psycopg2
django-bootstrap-form==3.2.1
